# resourceflow-cli/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ResourceFlowApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/admin/api-keys": {
      "get": {
        "tags": [
          "ApiKeys"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ApiKeys"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/api-keys/{id}/revoke": {
      "post": {
        "tags": [
          "ApiKeys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyScopeDto": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string"
          },
          "access": {
            "type": "string"
          }
        },
        "description": "One \u0060{resource, access}\u0060 pair \u2014 see ApiKeyScopes\nfor the allow-list. Doubles as the shape persisted in \u0060AdminApiKey.ScopesJson\u0060, so the\ncodec lives here rather than in the service: the read side (the authentication handler) and\nthe write side (the management service) both need it and neither owns the other."
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyScopeDto"
            }
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "An explicit expiry. Omit this to get the safe default (see\n            bool CreateApiKeyRequest.NeverExpires); when set, it must be in the future and\n            bool CreateApiKeyRequest.NeverExpires must not also be true.",
            "format": "date-time"
          },
          "neverExpires": {
            "type": "boolean",
            "description": "Opts a key out of the default one-year expiry entirely. Only meaningful when\n            DateTime? CreateApiKeyRequest.ExpiresAt is omitted \u2014 setting both is rejected rather than silently\n            preferring one, since a client sending both almost certainly means only one of them."
          }
        }
      },
...
    }
  },
...
}